rename matchString to match
