+++ test/matcher/type/MatchMatcher.spec.ts
@@ -1,10 +1,10 @@
 import {Matcher} from "../../../src/matcher/type/Matcher";
-import {matchString, not} from "../../../src/ts-mockito";
+import {match, not} from "../../../src/ts-mockito";
 
-describe("MatchingStringMatcher", () => {
+describe("MatchingMatcher", () => {
     describe("checking if value matches given regexp", () => {
-        const testObj: Matcher = matchString(/\w123/);
-        const notTestObj: Matcher = not().matchString(/\w123/);
+        const testObj: Matcher = match(/\w123/);
+        const notTestObj: Matcher = not().match(/\w123/);
 
         describe("when given value matches regexp", () => {
             it("returns true", () => {
